requirements.txt
torchvision==0.21.0
pyttsx3==2.98
opencv-python==4.11.0.86
torch==2.6.0
numpy==2.2.3
pandas==2.2.3
matplotlib==3.10.1
flask==3.1.0
timm==1.0.15
python-socketio==5.12.1
flask-socketio==5.5.1
eventlet==0.39.1
pillow==11.1.0
flask-sqlalchemy
bcrypt